# src/main/java/com/example/parceltracker/model/TrackingEvent.java
package com.example.parceltracker.model;

import java.time.Instant;

public record TrackingEvent(
        Instant timestamp,
        String rawStatus,
        String description,
        String location
) {
}